+++ src/Schema/Directives/WhereJsonContainsDirective.php
@@ -22,12 +22,6 @@ directive @whereJsonContains(
 GRAPHQL;
     }
 
-    /**
-     * Add a "WHERE JSON_CONTAINS()" clause to the builder.
-     *
-     * @param  \Illuminate\Database\Query\Builder|\Illuminate\Database\Eloquent\Builder  $builder
-     * @return \Illuminate\Database\Query\Builder|\Illuminate\Database\Eloquent\Builder
-     */
     public function handleBuilder($builder, $value): object
     {
         return $builder->whereJsonContains(