+++ src/Schema/Directives/WhereNotBetweenDirective.php
@@ -25,12 +25,6 @@ directive @whereNotBetween(
 GRAPHQL;
     }
 
-    /**
-     * Apply a "WHERE NOT BETWEEN" clause.
-     *
-     * @param  \Illuminate\Database\Query\Builder|\Illuminate\Database\Eloquent\Builder  $builder
-     * @return \Illuminate\Database\Query\Builder|\Illuminate\Database\Eloquent\Builder
-     */
     public function handleBuilder($builder, $values): object
     {
         return $builder->whereNotBetween(